Performance improvements
Use a number-keyed map instead of a JS object to avoid overhead of stringifying and hashing numbers.

Since we only need a single lowest cost item, find that, instead of doing a full array sort.

At a rough estimate, this gives a 25% speedup for comparing a data structure containing 700-element arrays.