Replace full diff with fast-deep-equal
Instead of doing a full diff, with pointer manipulation, just to check if two items are equal, this lets it exit out at the first inequality (assuming the code isn't using a custom diff implementation). This makes fast-deep-equal roughly 35 times faster for my test cases.  This is, obviously, a more invasive change: it adds a dependency and breaks the UMD build.